+++ src/union.js
@@ -36,12 +36,15 @@
 
 function union(Ts, name) {
 
+  assert(Arr.is(Ts) && Ts.every(isType), 'bad types');
+
   name = name || format('union(%s)', Ts.map(getName).join(', '));
 
   function Union(value, mut) {
     forbidNewOperator(this, Union);
     assert(Func.is(Union.dispatch), 'unimplemented %s.dispatch()', name);
     var T = Union.dispatch(value);
+    // a union type is idempotent iif every T in Ts is idempotent
     return T(value, mut);
   }