+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Roadmap
 =======
 
+# v.0.0.23
+
+- TODO create constants for string messages, fix #16
+
 # v.0.0.11
 
 - forbid the use of `new` for all types but structs, fix #8
@@ -9,12 +13,11 @@ Roadmap
 - make list(T) really idempotent, fix #11
 - make struct(T) really idempotent, fix #12
 - make tuple(Ts) really idempotent, fix #13
-- TODO make union(Ts) really idempotent, fix #14
+- make union(Ts) really idempotent, fix #14
 - export fail(), fix #15
-- TODO create constants for string messages, fix #16
 - TODO more tests on Dat primitive type, fix #18
 - TODO more tests on Re primitive type, fix #17
-- TODO make func() idempotent, fix #19
+- make func() idempotent, fix #19
 - relaxed `Obj.is()` to accept more object types
 - turn off jshint newcap
 - remove `freeze()` function